Use static getter for `ObjectBase::object_builder`. Using the static getter avoids the need to redeclare the (already static) variable in every function that wishes to use it.
Note that the object_builder is explicitly returning a reference, so that multiple calls to the function will always return the same instance. This avoids recurring issues that were previously coming up with `sqlite::database_binder::operator>>` not taking parameters by reference and causing side effects of function calls to be lost in ephemerality.