Use std::map with a complex key for `2log(n)` querying performance. By keying the map with by (trip, departure time) and having a custom comparator that sorts by trip then departure time, results for `next_visits` and `last_visits` can be found in `2log(n)` time.
With the custom comparator, Visits are stored in order, so a single search will return an iterator that can be used to find the next `n` visits on a trip. Simply filtering the iterator will return the next `n` visits to a given station. Using two queries will give a range of visits between two timepoints.