CSV Parsing overhaul. Timetable no longer relies on a CSV parsing library to read GTFS archives. Instead, it manually parses entries using `getline` with a comma delimiter and special-cases quoted string entries.
The decision to manually parse files was made because GTFS is a loose standard, meaning it has many optional fields that change the parsing requirements (number of columns, column order, cell formatting).

Getting the manual parsing to work also implied some minor changes to `field_mapper`, but these are negligible, as their only use case is in CSV parsing.