Prevent non-matching columns from being applied through `field_mapper`. This was far too difficult too discover for what the solution was, and was a problem born out of unnecessary complexity. So:
There was a typo in `stop_time.h` on the `drop_off_type` field mapping (was `dropoff_type`). This typo was exposing some unexpected behavior with `field_mapper`, and particularly with `type_map_t`. The best explanation is probably to walk through the steps of what was happening:

  1. `csv_parser::parse` determines the fields that are present in the CSV file, storing them in an ordered vector called `headers`.
  2. It then proceeds to read line by line, tokenizing each line and creating a vector parallel to `headers` called `columns`.
  3. For each pair in `(header, columns)` pair:
    a. The appropriate `field_mapper` instance (called `mapper`) is looked up based on `header`.
    b. Switching on the type that `mapper` represents, `column` is cast to that type and passed to `mapper.apply`.

The problem that came up starts becoming apparent through the details of how 3a works. Because the field `drop_off_type` did not exist in the parser's `field_types` (the typo had it as `dropoff_type`), and because `type_map_t` is an alias for `std::map<std::string, field_mapper>`, the behavior is to return a new instance of the map's `value_type` (in this case, `field_mapper`) using the default constructor. This default constructor does not set any of the members of `field_mapper`, and so their values are undefined.

Moving on to 3b, since the `type` member of this new `field_mapper` instance is undefined, it will almost always be left to match the `default` case of the switch statement. And here lay the second problem. The `default` case was placed above `case tSTRING`, as if to say "any columns without a type should be interpreted as strings". This may have been well intentioned, but it is nonsensical, as all `field_mapper` instances are automatically assigned a value for the `type` member.

Now, the cascading nature of switch statements takes effect and `mapper.apply(inst, string)` is called. However, since `string_target` has an undefined value, the memory address to which data is going to be written is also undefined. Experiential data showed that it was almost always the base address of `inst` (i.e., `string_target` was `NULL`, so the offset was 0), meaning the value would be written to the beginning of the memory for `inst`, regardless of the data that was actually stored there. In this case, that meant the value for `drop_off_type` was written over the data location for `trip_id` (the first data member of `stop_time`).

Because C++ is compiled and does very little in terms of dynamic data integrity checking, Timetable still saw `trip_id` as a valid string (even though the first few bytes of it were overwritten by the rogue `field_mapper` from above). Examining `inst.trip_id` would then give garbage values, as it had been corrupted. Even though most of the original data was there, some critical portion had been overwritten and the observed value was incorrect.

The solution for this whole issue was to simply move the `default` case of the switch statement in 3b to *after* all of the valid configurations, essentially making it a nop. The result (sufficiently) enforces that `header` map to an existing `field_mapper` instance in `field_types` in order to actually be applied.