Start `visits_after` implementation This initial implementation takes in parameters and returns a list of departure times (no arrival times, yet) matching those parameters.
Basic testing shows that the implementation is not accurate (visits are not returned in ascending time order), and it seems that the current structure of `visit_list` is partly at fault. The keying strategy currently keeps all visits on the same trip in ascending time order, but visits on the next trip may occur chronologically before them, resulting in an out-of-order result set, and thus incorrect results.