(#5, #6) Switch testing RPC to `visits_between` and adjust argument order. `visits_between` seems to be the call that will be made most commonly, so testing with it is more accurate to real-world use.
Per #5, the argument order of the RPCs has been adjusted to have `route` as the next-to-last parameter. At the moment, only the route-agnostic RPC is implemented.

Per #6, `route` and a temporary `headsign` placeholder have been added to each visit in an RPC response. I'm not entirely sure that `headsign` is a rational field to include, as it is an optional GTFS field, and doesn't really provide any information that can't be gleaned from the route itself.