(#16) Initial work on index system This commit introduces a method by which `stop_time`s are recorded and indexed to better serve various RPCs. Specifically, this system makes it easy to add new indices and more efficiently iterate `stop_time`s based on the needs of each RPC.
The indexing system works by creating maps with a custom indexing function (passed in as a lambdas to the `add_index` method of `Timetable::Timetable`). These functions are required to create unique identifiers to ensure that all `stop_time`s are represented in the index. To guarantee that an indexing function is capable of this uniqueness, it is passed a stop_time id parameter (the index of the stop_time in `Timetable`'s stop_time vector), which shouuld be appended to every generated key.

Indices are stored in a map, and are accessible via a string key through the `st_indices` member of `Timetable`. For example, an index with the name `station.departure` can be accessed via `tt.st_indices["station.departure"]`. As a `std::map`, all functions of maps and iterators around them are available.

Importantly, the map that indices are based on stores a `gtfs::stop_time*`, meaning that any modifications to the entries in the index must be done in-place. Should an element be moved in memory, all indices may be invalidated and require rebuilding. Luckily, this is a fairly efficient operation, but should generally be avoided.