Slight refactoring of `visits_between` for logical clarity. I'm still not satisfied with the amount of logic contained in `do_visits_between`, but I haven't found a good refactoring other than what has been done here.
My primary issue is with the duplication of logic in the primary function passed to WAMP to register RPCs. It would be nice to have some abstraction that takes care of logging requests and responses so that the functionality is always consistent across RPCs. However, I believe that the only real way to reach that point is to modify `wamped` (or begin a new implementation that better suits our needs).