Implement remaining `visits_*` RPCs. The entire `visits_*` family of RPCs should now be covered with proper implementations.
`visits_before` has a bit of a gotcha in the implementation, where it attempts to find the first visit immediately after those for the requested station in the `(route.)station.departure` indices, then decrements the resulting iterator by 1 to get the last visit at that station. It would be nice to have a method of entry to that visit that doesn't require the decrement, but `upper_bound` does not work as expected here, as simply requesting `upper_bound` with a station returns the first station in the index (it is equivalent to `lower_bound` in most cases).

Perhaps adding a fake departure time to the call to `upper_bound` would suffice. For example, using the string "ZZZ:ZZ:ZZ", which would lexicographically compare greater than any departure time that is less than 1000:00:00 (an absurd value that should never be met). Regardless, a cleaner method of entry would be nice.