Features and bugfixes to support BART and SFMTA (#22)
* Add a post_init member function to each gtfs class.

post_init is called after a gtfs entity has been parsed, before it is returned
from the csv_parser. This allows for custom setup logic in classes like `stop`
and `route`, which use post_init to handle gtfs sources that do not populate
fields like stop_code and route_short_name.

* Support parsing weekday in datetime.

* Handle CRLF strings in csv_parser.

* Print registration messages from transport.

* Specify transport and gtfs source as runtime arguments.

Previously, a static Timetable::Timetable instance was kept to share state
between components of the application. Since this instance now has to be
initialized from main(), I took the easy way out and made the static reference
a pointer, which required changing a bunch of `tt.` to `tt->`.

* Bug fixes for BART: stop_time referencing, DST calculation, route description.

* Fixed stop_time bug in every handler

* Made Visit's constructor take a Timetable pointer, for somewhat-nicer call site semantics.

* No need for std::placeholders in main.

* Use _getline with a non-default delimeter.

* Use _getline in a couple more places in csv_parser

* Chomp leading whitespace in the tokenizer