Fix csv parser buffer overrun on lines with a trailing comma.
(I originally wrote this as an issue, but fixed it before I could open it.)
Given a csv like...

     route_id,agency_id,route_short_name,route_long_name,route_desc,route_type,route_url,route_color,route_text_color
     01,BART,,Pittsburg/Bay Point - SFIA/Millbrae,,1,http://www.bart.gov/schedules/bylineresults?route=1,ffff33,

The header line is recognized as having 9 columns (as expected). The data line,
however, is recognized as having 8 entires, because the trailing `,` causes the
termination of the line parser. Rows are allocated as they are parsed, but the
header line is used to access values in the row. Because of this, this case
causes a buffer overrun. Interestingly, it *sometimes* works in macOS
(presumably why this bug leaked in) but so far always fails in linux. Some
kinda buffer size magic.

This change fixes the csv parser. I ran valgrind to verify that I could detect
the overrun and that this change fixes it.